Feature: Add functionality that creates code snippets based on the current line of code.
This commit introduces a number of functions that return a snippet of code based given the current line of code. It predicts what the next lines of code could be. The intention is that this functionality will save the developer time by not having to manually add lines of text.