Refactor: Improve performance of 'Smart rename symbol' handler.
Before, the `getNounInformation` function was invoked for each location of a symbol. However, that's unnecessary as the noun of the symbols is always the same regardles of the location of the symbol.

Now, the `getNounInformation` function is invoked before the loop that iterates through the list of a locations of symbol.