Fix: Solve bug with renaming a symbol that finishes with a number.
Before, the `Array.prototype.sort` method has a callback function. The parameters of the call back function usually end with a number. For example, `task1` and `task2` in `tasks.sort((task1, task2) => {})`. Those parameters couldn't be renamed accurately because they ended with a number. Now the already existing regular expression that matches the parameters of the callback function has been modified accordingly.