Feat: Get text selection to suggest code refactorings
Before, code refactorings were only considered for 1 line of code at most. This line of code was the one the caret was currently on.

Now, code refactorings are also suggested for text selections. For example, code refactoring of a `for... of` loop to `forEach` method where it's necessary to specify the boundaries of the `for... of` loop.