Refactor: Get server to handle "resource not found" better.
Before, the server required the existence of a `404.html` file everytime a resource was requested and not found. This also required the creation of a html file which added unnecessary clutter to the repository.

Now, the server simply sends `Not found!` as the body of the response. Otherwise, the server sends the requested resource.