Refactor: Remove workaround for a favicon
Before, the server would stopped code execution because the browser requested a favicon and the server couldn't find it. For that reason, a `link` element was added to the `tests.html` file as a workaround. This added unnecessary clutter to the html file.

Now, the `link` element has been removed since the server responds with a `resource not found` message instead of stopping execution.