Refactor: Updade algorithm to get symbols from a line of code.
Before, there were many regular expressions that matched symbols in a line of code. This was based on different patterns. For example, a `for of` loop, a `const` or `let` declaration, a `sort` method, etcetera. This list of regular expressions was intended to grow which would become a little difficult to manage.

Now, there is only 1 regular expression that matches symbols in a line of code. Sometimes, those symbols can be reserved keywords such as `for`, `in`, `debugger` and , thefore, they are filtered out.