Refactor: Simplify unit tests for code actions provider
Before, the unit tests relied on the very specific position of code snippets in an array of code snippets. For example, it was assumed that the code snippet at index 0 was always `Insert for... of loop`. However, the order is completely irrelevant and rendered the unit tests brittle.

Now, there isn't an array of code snippets and instead there's an object with keys. Each of those keys has a reference to a code snippet.