Refactor: Simplify unit tests for code actions provider
Before, there was some unnecessary repetition of code to test code actions for various lines of code. However, it became evident that lines of code written in different languages can result in the same code action. For example, the generation of a for... of loops for a line of code written in JavaScript or TypeScript.

Now, several lines of code can be added to an array of lines so that each one of them can be tested in a more straightforward manner.