Feat: Create code action to insert a code snippet that has a `forEach` method.
Before, it was ideal to use code action that inserted a code snippet that had a `for... of loop`. However, sometimes it was necessary to know the index of the element in an iterator.

Now, there's a new code action that inserts a code snippet that has a `forEach` method. There's an `index` parameter in the associated callback function.