Feat: Command can retrieve JavaScript file from anywhere in a workspace
Before, the `copyAsJavaScript` command assumed that compiled JavaScript files were always located in a `dist` folder. However, when this isn't the case, the command can't locate the equivalent JavaScript file.

Now, the command refers to the contents of the `tsconfig.json` file to find out where exactly JavaScript files are located.